Simply script for beefy checkpoint generation
